backend/requirements.txt
fastapi
uvicorn
pydantic
pillow
python-dotenv
google-generativeai
requests
replicate

